Fix crash
Previously, if the Handler did not invoke WriteHeader or Write on the
ResponseWriter, we would (a) never actually set up the header info and
(b) never set up the body object, so when we try to Close() it after the
call, we'd get a panic.

This has now been fixed; if the status has still not been set when the
Handler returns, we call WriteHeader(200) ourselves.