First pass and sandstorm-http-bridge support
The websession stuff will probably take long enough to get going that
we'll want this.

For now this doesn't work, because sandstorm-http-bridge doesn't start
listening for connections until it detects an http server. This has been
a TODO for sandstorm for ~2.5 years, and since it makes our API awkward
to accommodate it, I'm going to just try to fix that myself.