First whack at postStreaming.
This involved factoring out some of the bits from handleCommon

The test currently fails, because the http client blocks reading from
the request body; it never even calls postStreaming. This is me not
understanding how you're supposed to do streaming request bodies
client-side. I need to figure this out, but I'm calling it a day.

Once this is working, putStreaming will be trivial.