Use capnp in session cookies, fix panic
The panic was because we were looking for a session id, which is
independent of login status; instead, look to see if the credentials are
there. This avoids the code path that was previously panicking.

It looks like data still isn't getting saved? After login we get told
we're still not logged in; need to investigate.

This also moves over to using capnp to store data in the session cookie,
so we can use the same evolveability strategy here as everywhere else.