Misc. work on frontend.
We now have links that try to open iframes, plush some misc. stubs.

I say "try" since the server end of this isn't ready yet -- also we'll
need to futz with this to make sure we don't reload the app page when we
switch between apps, but I think I know how that's going to work.