Correctly shut down grains when exiting.
There's a lot here. The high-level is that we want to kill the grain's
pid namespace's root, and then have the sandbox-launcher sitting outside
of it wait on the grain, while we wait for the launcher. This extra
process only really exists because we didn't want to write the logic for
it in Go.

To make this work, we need to know the external pid of the grain's root
process, so the launcher now writes that to fd #4, and the main tempest
process reads it back.

The patch also adds a bunch of logging.

Finally, we actually need to run the shutdown procedure, and calling
os.Exit() in Main() was preventing the deferred srv.Release() from
executing. So we call panic instead.