ui: Defunctionalize ui messages.
I should write a proper architecture document, but the idea here is to
make this a bit more declarative by sending data instead of just a
lambda. This way it's also possible to just look at message.go to find
all the possible events that the UI might react to.