Update use of streaming APIs for go-capnp change.
Since https://github.com/capnproto/go-capnproto2/pull/450 has been
merged, the streaming APIs have changed; this patch updates uses of
ByteStream and WebSocketStream to match.

In the course of doing this I noticed we were not checking errors or
releasing calls in some of the ByteStream code, so this is now also
fixed.