Reorganize directory hierarchy, add make check target.
This moves the database and grain-agent directories into server/.
Orignally I'd left these separate, planning on only putting things in
server/ which were not used outside of the server binary. But this makes
it annoying to do things like test all of the server-side code.

This patch also adds a `check` target to the Makefile, which runs the
tests.