Manifest.nounPhrase: default to "instance."
This solves the problem that some older apps don't set this. Sandstorm
jumps through hoops to derive a sensible value from title if this is not
found, but F that noise. There are like 3 apps we care about that do
this, and we can fix the manifests at are leisure.

This means that e.g. Davros now says "instance" instead of "shared
directory" which is not so cool, but like I said, fix the app.

This is technically a breaking schema change, but it is unlikely to
actually break anything.